Permission v8

+++ app/src/main/java/org/xksyu/mca/feature/basic/AlarmReceiver.kt
@@ -26,6 +26,7 @@ import java.util.Calendar
 class AlarmReceiver : BroadcastReceiver() {
     @SuppressLint("ServiceCast")
     override fun onReceive(context: Context, intent: Intent) {
+        val settingsManager = SettingsManager(context)
         when(intent.action){
             "ACTION_ALARM_GET" -> {
                 val alarmId = intent.getIntExtra("ALARM_ID", -1)
@@ -39,48 +40,54 @@ class AlarmReceiver : BroadcastReceiver() {
             }
             Intent.ACTION_BOOT_COMPLETED -> {
                 CoroutineScope(Dispatchers.Main).launch{
-                    reloadList(context)
+                    reloadList(context, settingsManager)
                 }
             }
             "STOP_ALARM_ACTION" -> {
-                val alarmId = intent.getIntExtra("ALARM_ID", -1)
-                val notificationManager = context.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
-                val serviceIntent = Intent(context, AlarmForegroundService::class.java)
-                context.stopService(serviceIntent)
-                val repository = (context.applicationContext as MCApplication).repository
-                CoroutineScope(Dispatchers.IO).launch {
-                    val alarm = repository.getById(alarmId)
-                    alarm?.let {
-                        if (it.isRepeat) {
-                            repository.deleteAlarm(it)
+                if(!settingsManager.debugGet()) {
+                    val alarmId = intent.getIntExtra("ALARM_ID", -1)
+                    val notificationManager =
+                        context.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
+                    val serviceIntent = Intent(context, AlarmForegroundService::class.java)
+                    context.stopService(serviceIntent)
+                    val repository = (context.applicationContext as MCApplication).repository
+                    CoroutineScope(Dispatchers.IO).launch {
+                        val alarm = repository.getById(alarmId)
+                        alarm?.let {
+                            if (it.isRepeat) {
+                                repository.deleteAlarm(it)
+                            }
+                            println("NOTICE  STOP   id: ${it.id}")
                         }
-                        println("NOTICE  STOP   id: ${it.id}")
+                        notificationManager.cancel(alarmId)
                     }
-                    notificationManager.cancel(alarmId)
                 }
             }
             "SNOOZE_ALARM_ACTION" -> {
-                val notificationManager = context.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
-                val alarmId = intent.getIntExtra("ALARM_ID", -1)
-                val settingsManager = SettingsManager(context)
-                val repository = (context.applicationContext as MCApplication).repository
-                CoroutineScope(Dispatchers.IO).launch {
-                    val alarm = repository.getById(alarmId)
-                    alarm?.let{
-                        val repeatAlarm = it.copy(
-                            id = settingsManager.updateId(),
-                            isRepeat = true, autoWeek = 0
-                        )
-                        setAlarm(repeatAlarm,context)
-                        repository.insertAlarm(repeatAlarm)
-                        if (it.isRepeat){
-                            repository.deleteAlarm(it)
-                            println("NOTICE  AGAIN   id: ${it.id}")
+                if(!settingsManager.debugGet()) {
+                    val notificationManager =
+                        context.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
+                    val alarmId = intent.getIntExtra("ALARM_ID", -1)
+                    val settingsManager = SettingsManager(context)
+                    val repository = (context.applicationContext as MCApplication).repository
+                    CoroutineScope(Dispatchers.IO).launch {
+                        val alarm = repository.getById(alarmId)
+                        alarm?.let {
+                            val repeatAlarm = it.copy(
+                                id = settingsManager.updateId(),
+                                isRepeat = true, autoWeek = 0
+                            )
+                            setAlarm(repeatAlarm, context, settingsManager)
+                            repository.insertAlarm(repeatAlarm)
+                            if (it.isRepeat) {
+                                repository.deleteAlarm(it)
+                                println("NOTICE  AGAIN   id: ${it.id}")
+                            }
                         }
+                        val serviceIntent = Intent(context, AlarmForegroundService::class.java)
+                        context.stopService(serviceIntent)
+                        notificationManager.cancel(alarmId)
                     }
-                    val serviceIntent = Intent(context, AlarmForegroundService::class.java)
-                    context.stopService(serviceIntent)
-                    notificationManager.cancel(alarmId)
                 }
             }
             "DELETE_ACTION" -> {
@@ -185,142 +192,10 @@ class AlarmForegroundService : Service() {
     override fun onBind(intent: Intent?): IBinder? = null
 }
 
-fun setAlarm(alarm: AlarmData, context: Context) {
-    println("----------------setAlarm---------------")
-
-    var dayOfWeek = 0b1 shl (Calendar.getInstance().get(Calendar.DAY_OF_WEEK) - 2)
-    if (dayOfWeek == 0) dayOfWeek = 0b1 shl 6
-    val time = when(alarm.autoWeek) {
-        3 -> {
-            Calendar.getInstance().apply {
-                set(Calendar.HOUR_OF_DAY, alarm.timeHour)
-                set(Calendar.MINUTE, alarm.timeMinute)
-                set(Calendar.SECOND, 0)
-                set(Calendar.MILLISECOND, 0)
-                if (timeInMillis <= System.currentTimeMillis()){
-                    add(Calendar.DATE,1)
-                }
-            }
-        }
-
-        2 -> {
-            println("  dayOfWeek: ${dayOfWeek.toString(2)}")
-            Calendar.getInstance().apply {
-                set(Calendar.HOUR_OF_DAY, alarm.timeHour)
-                set(Calendar.MINUTE, alarm.timeMinute)
-                set(Calendar.SECOND, 0)
-                set(Calendar.MILLISECOND, 0)
-                val set: Int = when (alarm.weekSelect) {
-                    0b1 -> 0b00_11111_00_11111
-                    0b10 -> 0b11_00000_11_00000
-                    else -> 0b11_11111_11_11111
-                }
-                println("  set: ${set.toString(2)}")
-                for (count in 0..7) {
-                    if (((dayOfWeek shl count) and set) != 0) {
-                        println("  value: ${((dayOfWeek shl count) and set).toString(2)}, count: $count")
-                        add(Calendar.DATE, count)
-                        if (timeInMillis <= System.currentTimeMillis()) continue
-                        break
-                    }
-                }
-            }
-        }
-
-        1 -> {
-            val setTemp = alarm.weekSelect
-            val set = (setTemp shl 7) + alarm.weekSelect
-            println("  dayOfWeek: ${dayOfWeek.toString(2)}, setTemp: ${setTemp.toString(2)}, set: ${set.toString(2)}")
-            Calendar.getInstance().apply {
-                set(Calendar.HOUR_OF_DAY, alarm.timeHour)
-                set(Calendar.MINUTE, alarm.timeMinute)
-                set(Calendar.SECOND, 0)
-                set(Calendar.MILLISECOND, 0)
-                println("DEBUG week set")
-                for (count in 0..7) {
-                    if (((dayOfWeek shl count) and set) != 0) {
-                        println("  value: ${((dayOfWeek shl count) and set).toString(2)}, count: $count")
-                        add(Calendar.DATE, count)
-                        if (timeInMillis <= System.currentTimeMillis()) continue
-                        break
-                    }
-                }
-            }
-        }
-
-        else -> {
-            Calendar.getInstance().apply {
-                set(Calendar.HOUR_OF_DAY, alarm.timeHour)
-                set(Calendar.MINUTE, alarm.timeMinute)
-                set(Calendar.SECOND, 0)
-                set(Calendar.MILLISECOND, 0)
-                if(alarm.isRepeat){
-                    add(Calendar.MINUTE,alarm.remindMinute)
-
-                    //
-                    //add(Calendar.SECOND,10)
-                    //add(Calendar.MINUTE, - alarm.remindMinute)
-                    //For debug
-
-                    alarm.timeMinute += alarm.remindMinute
-                    if (alarm.timeMinute>=60){
-                        alarm.timeMinute -= 60
-                        alarm.timeHour += 1
-                        if (alarm.timeHour>=24) alarm.timeHour-=24
-                    }
-                }else if (timeInMillis <= System.currentTimeMillis()){
-                    add(Calendar.DATE,7)
-                }
-            }
-        }
-    }
-
-    println("DEBUG AlarmData values:")
-    println("  id: ${alarm.id}")
-    println("  timeHour: ${alarm.timeHour}, timeMinute: ${alarm.timeMinute}")
-    println("  name: ${alarm.name}")
-    println("DEBUG Set values:")
-    println("  DayOfWeek: ${time.get(Calendar.DAY_OF_WEEK)}")
-    println("  Hour: ${time.get(Calendar.HOUR_OF_DAY)}, Minute: ${time.get(Calendar.MINUTE)}")
-
-    val alarmIntent = Intent(context, AlarmReceiver::class.java).apply {
-        action = "ACTION_ALARM_GET"
-        putExtra("ALARM_ID", alarm.id)
-    }
-    val alarmPendingIntent = PendingIntent.getBroadcast(
-        context,
-        alarm.id,
-        alarmIntent,
-        PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_IMMUTABLE
-    )
-
-    val alarmManager = context.getSystemService(Context.ALARM_SERVICE) as AlarmManager
-    alarmManager.setExactAndAllowWhileIdle(AlarmManager.RTC_WAKEUP,time.timeInMillis,alarmPendingIntent)
-}
-
-fun cancelAlarm(alarm: AlarmData, context: Context){
-    println("====== Call cancelAlarm BEGIN ======")
-    val alarmIntent = Intent(context, AlarmReceiver::class.java).apply {
-        action = "ACTION_ALARM_GET"
-        putExtra("ALARM_ID", alarm.id)
-    }
-    val alarmPendingIntent = PendingIntent.getBroadcast(
-        context,
-        alarm.id,
-        alarmIntent,
-        PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_IMMUTABLE
-    )
-
-    val alarmManager = context.getSystemService(Context.ALARM_SERVICE) as AlarmManager
-    alarmManager.cancel(alarmPendingIntent)
-    alarmPendingIntent.cancel()
-    println("====== Call cancelAlarm FINISHED ======")
-}
-
-suspend fun reloadList(context: Context){
+suspend fun reloadList(context: Context,settingsManager: SettingsManager){
     val repository = (context.applicationContext as MCApplication).repository
     val list = repository.alarms.first()
     list.forEach { alarm ->
-        setAlarm(alarm, context)
+        setAlarm(alarm, context, settingsManager)
     }
 }
